Enhance versioning and update dependencies
Updated the versioning logic in the `nuget-main.yml` workflow to skip non-packable C# projects by checking the `IsPackable` property. Upgraded the .NET version in the workflow from `9.0.x` to `10.0.x`. Updated the Visual Studio version in the solution file to `18.0.11205.157`. Added the `nuget-main.yml` file to the `Pub` solution folder.